separate d3 visualization to an imported  module
